Move valid_available_subdomain to validators.py
Update the signature of valid_available_subdomain to be a field
validator, raising an exception on failure and returning None.
