+++ emails/tests/models_tests.py
@@ -25,16 +25,11 @@ from ..models import (
     DeletedAddress,
     DomainAddress,
     Profile,
-    RegisteredSubdomain,
     RelayAddress,
     address_hash,
     get_domain_numerical,
-    has_bad_words,
-    hash_subdomain,
-    is_blocklisted,
     valid_address,
     valid_address_pattern,
-    valid_available_subdomain,
 )
 from ..utils import get_domains_from_settings
 
@@ -127,44 +122,7 @@ def vpn_subscription() -> str:
     return random.choice(vpn_only_plans)
 
 
-class MiscEmailModelsTest(TestCase):
-    def test_has_bad_words_with_bad_words(self):
-        assert has_bad_words("angry")
-
-    def test_has_bad_words_without_bad_words(self):
-        assert not has_bad_words("happy")
-
-    def test_has_bad_words_exact_match_on_small_words(self):
-        assert has_bad_words("ho")
-        assert not has_bad_words("horse")
-        assert has_bad_words("ass")
-        assert not has_bad_words("cassandra")
-        assert has_bad_words("hell")
-        assert not has_bad_words("shell")
-        assert has_bad_words("bra")
-        assert not has_bad_words("brain")
-        assert has_bad_words("fart")
-        assert not has_bad_words("farther")
-        assert has_bad_words("fu")
-        assert not has_bad_words("funny")
-        assert has_bad_words("poo")
-        assert not has_bad_words("pools")
-
-    def test_is_blocklisted_with_blocked_word(self):
-        assert is_blocklisted("mozilla")
-
-    def test_is_blocklisted_with_custom_blocked_word(self):
-        # custom blocked word
-        # see MPP-2077 for more details
-        assert is_blocklisted("customdomain")
-
-    def test_is_blocklisted_without_blocked_words(self):
-        assert not is_blocklisted("non-blocked-word")
-
-    @patch("emails.models.emails_config", return_value=Mock(blocklist=["blocked-word"]))
-    def test_is_blocklisted_with_mocked_blocked_words(self, mock_config: Mock) -> None:
-        assert is_blocklisted("blocked-word")
-
+class AddressHashTest(TestCase):
     @override_settings(RELAY_FIREFOX_DOMAIN="firefox.com")
     def test_address_hash_without_subdomain_domain_firefox(self):
         address = "aaaaaaaaa"
@@ -191,10 +149,14 @@ class MiscEmailModelsTest(TestCase):
         expected_hash = sha256(f"{address}@{test_domain}".encode()).hexdigest()
         assert address_hash(address, domain=test_domain) == expected_hash
 
+
+class GetDomainNumericalTest(TestCase):
     def test_get_domain_numerical(self):
         assert get_domain_numerical("default.com") == 1
         assert get_domain_numerical("test.com") == 2
 
+
+class ValidAddressPatternTest(TestCase):
     def test_valid_address_pattern_is_valid(self):
         assert valid_address_pattern("foo")
         assert valid_address_pattern("foo-bar")
@@ -377,11 +339,11 @@ class RelayAddressTest(TestCase):
         relay_address.delete()
         assert not valid_address(relay_address.address, relay_address.domain_value)
 
-    @patch(
-        "emails.models.emails_config",
-        return_value=Mock(badwords=[], blocklist=["blocked-word"]),
-    )
-    def test_address_contains_blocklist_invalid(self, mock_config: Mock) -> None:
+    @patch("emails.validators.badwords", return_value=[])
+    @patch("emails.validators.blocklist", return_value=["blocked-word"])
+    def test_address_contains_blocklist_invalid(
+        self, mock_blocklist: Mock, mock_badwords: Mock
+    ) -> None:
         blocked_word = "blocked-word"
         relay_address = RelayAddress.objects.create(
             user=baker.make(User), address=blocked_word
@@ -975,82 +937,6 @@ class ProfileSaveTest(ProfileTestCase):
             assert relay_address.description == self.TEST_DESCRIPTION
             assert relay_address.generated_for == self.TEST_GENERATED_FOR
             assert relay_address.used_on == self.TEST_USED_ON
-
-
-class ValidAvailableSubdomainTest(TestCase):
-    """Tests for valid_available_subdomain()"""
-
-    ERR_NOT_AVAIL = "error-subdomain-not-available"
-    ERR_EMPTY_OR_NULL = "error-subdomain-cannot-be-empty-or-null"
-
-    def reserve_subdomain_for_new_user(self, subdomain: str) -> User:
-        user = make_premium_test_user()
-        user.profile.add_subdomain(subdomain)
-        return user
-
-    def test_bad_word_raises(self) -> None:
-        with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
-            valid_available_subdomain("angry")
-
-    def test_blocked_word_raises(self) -> None:
-        with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
-            valid_available_subdomain("mozilla")
-
-    def test_taken_subdomain_raises(self) -> None:
-        subdomain = "thisisfine"
-        self.reserve_subdomain_for_new_user(subdomain)
-        with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
-            valid_available_subdomain(subdomain)
-
-    def test_taken_subdomain_different_case_raises(self) -> None:
-        self.reserve_subdomain_for_new_user("thIsIsfInE")
-        with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
-            valid_available_subdomain("THiSiSFiNe")
-
-    def test_inactive_subdomain_raises(self) -> None:
-        """subdomains registered by now deleted profiles are not available."""
-        subdomain = "thisisfine"
-        user = self.reserve_subdomain_for_new_user(subdomain)
-        user.delete()
-
-        registered_subdomain_count = RegisteredSubdomain.objects.filter(
-            subdomain_hash=hash_subdomain(subdomain)
-        ).count()
-        assert Profile.objects.filter(subdomain=subdomain).count() == 0
-        assert registered_subdomain_count == 1
-        with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
-            valid_available_subdomain(subdomain)
-
-    def test_subdomain_with_space_raises(self) -> None:
-        with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
-            valid_available_subdomain("my domain")
-
-    def test_subdomain_with_special_char_raises(self) -> None:
-        with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
-            valid_available_subdomain("my@domain")
-
-    def test_subdomain_with_dash_returns_True(self) -> None:
-        assert valid_available_subdomain("my-domain") is True
-
-    def test_subdomain_with_dash_at_front_raises(self) -> None:
-        with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
-            valid_available_subdomain("-mydomain")
-
-    def test_empty_subdomain_raises(self) -> None:
-        with self.assertRaisesMessage(
-            CannotMakeSubdomainException, self.ERR_EMPTY_OR_NULL
-        ):
-            valid_available_subdomain("")
-
-    def test_null_subdomain_raises(self) -> None:
-        with self.assertRaisesMessage(
-            CannotMakeSubdomainException, self.ERR_EMPTY_OR_NULL
-        ):
-            valid_available_subdomain(None)
-
-    def test_subdomain_with_space_at_end_raises(self) -> None:
-        with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
-            valid_available_subdomain("mydomain ")
 
 
 class ProfileDisplayNameTest(ProfileTestCase):
